LayoutTests/platform/ios-simulator/editing/selection/click-in-focusable-link-should-not-clear-selection-expected.txt

Ensures that the selection is not cleared when the user does a mousedown on a mouse focusable link.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL String(window.getSelection()) should be Select. Was .
FAIL String(window.getSelection()) should be Select. Was .
PASS successfullyParsed is true

TEST COMPLETE
